Support sending chat feedback (#86)
Wires up sending chat feedback to our telemetry backend.

Feedback has to be sent to server for conversationId. Unfortunately, the only way right now we can receive the event back is via telemetry event from the server side. So a fake metric name amazonq_sendFeedback with the required data is emitted from the server which is then intercepted on the client side and sent to toolkit SIM feedback backend.