Add csp for login view and only check for overrides if atleast one variable is set (#179)
This change adds csp to the Toolkit login webview html. In addition this fixes a bug where we always validated override result for lsp even when none of the override env variables were set. We now do attempt to validate the overrides unless atleast one of the env variables are set.