Add support for using fallback lsp versions and deleting old version (#213)
Follow-up to a series of PRs( #115 #129 #162) that improves LSP fetching logic.
This change adds support for using a fallback version when the latest lsp version in the toolkit compatible range cannot be fetched locally or from a remote location. This allows using a valid locally cached old language server version as a fallback.
A lsp version is considered a valid fallback if it is:
* in the toolkit compatible version range
* is not delisted
* has a valid local cache i.e. hash of the locally cached version of the language server matches the hash for the specified version of language server in the version manifest
In addition, this also adds cleanup logic to remove old versions of the language server in the cache location. Cleanup helps keep the storage reasonably clean.
It is executed on a background thread asynchronously and is triggered after a 10sec delay after the Language server download process completes successfully.
It cleans up versions cached that are in the toolkit compatible version range. This includes:
*deleting any delisted versions in that range
* deleting all but the highest 2 versions in that range
Bonus: Minor fix to use the most recent version of language server instead of picking first from version manifest
The version range for language servers has been reduced to be between [0.0.0-1.0.0)