Error handling improvement for Q chat (#223)
Improvements to error messaging shown to users with Q chat related logic.
Refactored the error handling for showing error responses in Chat UI to include exceptions occurred when parsing the encrypted chat result or when formulating the encrypted request to the language server. This ensures users will see some response in UI if any step in the process of sending the request to the lsp and receiving a response fails.
In addition, some minor tweaks have been made to current logs and some not required logs have been removed.