Fix issue where ideApi browser function is injected before browser is fully loaded (#474)
This change fixes an issue where users get stuck on Amazon Q login screen and clicking continue does not result any action.
This was caused because the window.ideApi which connects the javascript to IDE was null. This was because of a similar issue highlighted here: eclipse-platform/eclipse.platform.swt#1274.
This issue follows the advice given here where the injection of the browserfunction now happens after the browser has fully loaded by calling in as part of the completed callback of the browser.progresslistener. This ensures the ideApi and telemetryApi callbacks are always there after load.
Browser setup now follows the given order:

* Create browser
* Load content with setText
* Wait for load to complete fully
* Inject browserFunctions