Pass IDE trusted system certs to Language server by default (#482)
This change improves the proxy support story for the extension. With this, we honor customer CA cert if specified in the preferences UI. If it is not supplied, instead of leaving it blank, we now detect system certificates and send it over to the node based language server. This allows us to address some issues where users are on corporate proxies/firewalls that have a proxy url but not an explicitly defined cert and expects applications to honor system certs.
We currently do the same system cert detection when downloading artifacts for lsp.

Follows a similar approach as JB: aws/aws-toolkit-jetbrains#5553